# Reduce UI blocking time when saving large notebooks by using file stream