--- a/packages/filebrowser/package.json
+++ b/packages/filebrowser/package.json
@@ -60,6 +60,7 @@
     "@lumino/signaling": "^2.1.4",
     "@lumino/virtualdom": "^2.0.3",
     "@lumino/widgets": "^2.7.1",
+    "jest-environment-jsdom": "^29.3.0",
     "react": "^18.2.0"
   },
   "devDependencies": {