--- a/packages/services/src/contents/index.ts
+++ b/packages/services/src/contents/index.ts
@@ -1792,29 +1792,6 @@ export class RestContentProvider implements IContentProvider {
   ): Promise<Contents.IModel> {
     const settings = this._options.serverSettings;
     const url = this._getUrl(localPath);
-    // requires http2 in tornado!
-    /*
-    function createJSONStreamFromString(str: string, chunkSize = 65536) {
-      let offset = 0;
-      return new ReadableStream({
-        pull(controller) {
-          if (offset >= str.length) {
-            controller.close();
-            return;
-          }
-          const chunk = str.slice(offset, offset + chunkSize);
-          controller.enqueue(new TextEncoder().encode(chunk));
-          offset += chunkSize;
-        }
-      });
-    }
-    const stream = createJSONStreamFromString(JSON.stringify(options));
-    const init = {
-      method: 'PUT',
-      body: stream,
-      duplex: 'half'
-    };
-    */
     const file = new File([JSON.stringify(options)], 'data.json', {
       type: 'application/json'
     });