--- a/packages/testing/src/jest-env.ts
+++ b/packages/testing/src/jest-env.ts
@@ -17,5 +17,7 @@ export default class FixJSDOMEnvironment extends JSDOMEnvironment {
     this.global.Headers = Headers;
     this.global.Request = Request;
     this.global.Response = Response;
+    // While jsdom has it's own File, it clashes with instance checks in fetch, so we use one from Node.js
+    this.global.File = File;
   }
 }