## References

- Alleviates https://github.com/jupyterlab/jupyterlab/issues/17017

## Code changes

Instead of sending the string via HTTP PUT request in one chunk, send the contents via File upload which allows the tornado server to receive it in chunks.

## User-facing changes

### UI blocking time

| Browser | Before | After | Speedup |
|--|--|--|--|
| Chrome | 580ms, 550ms, 564ms | 422ms, 409ms, 414ms | 26.5% |
| Firefox |  322ms, 322ms, 326ms | 133ms, 131ms, 133ms |59.0% |

<details>

For Chrome and Firefox I used a notebook of 98.7 MB with the result of printing out 10**7 lines as in the issue reproducer. This was too much for Safari/Webkit to render so instead I generated a single image output which weighs 100MB when saved with:

```python
import numpy as np
from PIL import Image
import io
import base64
from IPython.display import display, HTML

# Desired final output: ~100 MB base64 string
# So raw image should be ~75 MB
target_raw_bytes = int(75 * 1024 * 1024)  # 75 MB
channels = 3
bytes_per_pixel = 1  # uint8

# Compute image size (square image)
total_pixels = target_raw_bytes // (channels * bytes_per_pixel)
side = int(np.ceil(np.sqrt(total_pixels)))

# Generate random image (not compressible)
array = np.random.randint(0, 256, (side, side, channels), dtype=np.uint8)
img = Image.fromarray(array, 'RGB')

# Save as PNG (mild compression, but adds headers)
buf = io.BytesIO()
img.save(buf, format='PNG', compress_level=0)  # use lowest compression
buf.seek(0)

# Base64 encode
b64 = base64.b64encode(buf.read()).decode('utf-8')
data_uri = f'data:image/png;base64,{b64}'

# Display using HTML
html = f'<img src="{data_uri}" />'
display(HTML(html))
```

But I was not able to find a way to measure time take on Safari/WebKit. In any case this was tested on Safari  and GNOME Web too.

</details>

## Backwards-incompatible changes

None

Tested against `jupyter-server` v2 and against `jupyverse` 0.10.3.